spaceapi: Move code from lib.rs into status.rs
lib.rs now only contains references to the submodules
